Chunks: Move to namespace, fix nasty bug in IDAT
The IDAT chunk was continuously clobbering the previous scanline
when reading scanlines because I forgot that byte[]s are passed
by reference and not by value. Thus, the read-filter-write loop
in GetCompressedDataAsync would, at the beginning of every iteration,
nicely read the next scanline in, and because we'd foolishly
aliased `previousScanline` to `scanline`, we'd end up with both
scanlines always being the same. This didn't present an issue to the
Sub filter, but with the Up filter, it resulted in images that simply
repeated the first scanline, ad nauseam.