Filters: Swap API from Filter to FilterInto
The original API pretty much guaranteed that Filter had to do an
allocation, which the caller would then have to copy into its buffer
or stream. In our case, since the caller deals exclusively with byte[]
buffers, it makes more sense for it to allocate the buffer and pass
it to filter.