Benchmarks: Working benchmark suite!
Includes Adler32 and Sub implementations. The benchmark
suite currently works by having a "mirror" project.json
for the benchmarks, and in order to run it, you must switch
your .NET Core SDK to preview2. Once upstream updates, we'll
hopefully be able to get rid of that and use the .csproj.

Adler32 currently includes 4 implementations--a "smartest" that
is close to upstream zlib's implementation, a "smarter" that
does essentially what smartest does but without a number of unrolled
16-byte loop operations first (it simply goes byte-by-byte),
and two more naive implementations that go byte-by-byte without
avoiding modulo math.

Sub includes 3 implementations--a naive one with byte[]'s, a
slightly optimized one that uses byte*'s, and an incredibly optimized
implementation that uses vectors. On sizeable scanlines, this
implementation is 10-15x faster than either of the other two, when
run on RyuJIT. The legacy JIT and Mono are less good, so we may
need to add some JIT-detection code to decide which Sub implementation
we're going to use.