Filters: Implement optimized Sub filter
Update benchmarks with latest optimizations, and actually
bring the optimized implementation over to Pingu core. The
optimized filter selects between a vectorized implementation
on non-Mono platforms with hardware accelerated Vector<T>
(ie. RyuJIT), and an unrolled pointer loop on other platforms
(the runaway winner in benchmarks on Legacy and Mono JITs).

The implementation is selected at first use and a flag is set for
subsequent calls--the static readonly flag will not cause any performance
hit, as the incorrect branch will be eliminated by the JIT when
FilterInto is first called due to static readonly primitives being
treated like constants.