Filters: Implement Paeth filter
Implements the Paeth filter using an optimized implementation as follows:

When the previous scanline is null, fall back to the Sub filter's
implementation, as Paeth(Raw(x-bpp), 0, 0) = Raw(x-bpp), and that's
the very definition of Sub. Otherwise, use a fast absolute value
calculation, along with an unrolled loop and aggressive inlining
hinting for both the predictor and the absolute value calculation
for maximum speed.

Fixes #4.