WIP: Image sources: Pallete-indexed image source & conversions
All conversions are written except for indexed -> indexed conversion
that widens the bit depth. The grayscale w/ alpha conversions are
extremely well tested, the truecolor tests are starting up,
and the grayscale w/o alpha tests need to be written. The grayscale
tests will be extremely complex, as grayscale images can be in any bit
depth of 1, 2, 4, 8, and 16. This means we need 1 test for each source and
target bit depth, for a total of 20 tests just for indexed -> grayscale
conversion.